moving writeconfig to new file and adding tests
